Move round_y_axis_max to etl.utils and drop unused convert_numpy_types
Relocate round_y_axis_max (and its tests) from immunohisto_transform into
etl.utils so it can be shared across transforms, remove the unused
convert_numpy_types helper, and qualify utils references in test_utils.py.
