+++ src/agoradatatools/etl/transform/immunohisto_transform.py
@@ -4,13 +4,13 @@ This is for the Model AD project.
 """
 
 import pandas as pd
-import math
-from typing import Dict, List, Any, Union, Tuple
+from typing import Dict, List, Any, Tuple
 
 from agoradatatools.etl.utils import (
     check_required_datasets_and_columns,
     nest_fields,
     normalize_null_values,
+    round_y_axis_max,
 )
 
 
@@ -67,56 +67,6 @@ def prepare_immunohisto_data(df: pd.DataFrame) -> pd.DataFrame:
     df["age"] = df["age"].apply(lambda x: x if x.endswith("months") else x + " months")
 
     return df
-
-
-def round_y_axis_max(y_axis_max: Union[int, float, str]) -> float:
-    """
-    This function rounds the y_axis_max value to the nearest sensible nice round number.
-
-    Logic:
-    - If max == 0, then y_axis_max == 10
-    - Else, round UP to the next "nice" number where the second digit is 0 or 5
-
-    Examples:
-    - 0.0021 rounds up to 0.0025
-    - 0.0004 rounds up to 0.00045
-    - 0.329486078 rounds up to 0.35
-    - 0.089 rounds up to 0.090
-    - 1094 rounds up to 1500
-    - 1322498 rounds up to 1500000
-    - 728591 rounds up to 750000
-    - 3973 rounds up to 4000
-    - 1.616 rounds up to 2.0
-    """
-    # Convert to float if it's a string or other type
-    try:
-        y_axis_max = float(y_axis_max)
-    except (ValueError, TypeError):
-        # If conversion fails, return 10 (default case)
-        return 10.0
-
-    # Handle special cases: zero or negative values
-    if y_axis_max <= 0:
-        return 10.0 if y_axis_max == 0 else 0.0
-
-    # Find the order of magnitude of the number
-    magnitude = int(math.floor(math.log10(y_axis_max)))
-
-    # Scale the number so the first digit is in the ones place
-    scaled = y_axis_max / (
-        10 ** (magnitude - 1)
-    )  # After scaling, the first significant digit is in the tens place and the second in the ones place
-    # Round UP to the next 5 or 0 (add small epsilon to ensure we always round up)
-    rounded_scaled = 5 * math.ceil((scaled + 1e-10) / 5)
-
-    # Put back to the right magnitude
-    result = rounded_scaled * (10 ** (magnitude - 1))
-
-    # Remove float precision issues
-    result = round(result, 15)
-
-    # Ensure we always return a float, even for whole numbers
-    return float(result)
 
 
 def _calculate_y_axis_max_map(
