+++ tests/test_utils.py
@@ -12,16 +12,11 @@ import synapseclient
 import yaml
 
 from agoradatatools.etl import utils
-from agoradatatools.etl.utils import (
-    ColumnRule,
-    ContainsSubstringRule,
-    MatchesRegexRule,
-    NotEmptyRule,
-    OneOfRule,
-)
 
 RuleClassWithValueArg = (
-    type[MatchesRegexRule] | type[ContainsSubstringRule] | type[OneOfRule]
+    type[utils.MatchesRegexRule]
+    | type[utils.ContainsSubstringRule]
+    | type[utils.OneOfRule]
 )
 
 
@@ -751,30 +746,30 @@ class TestColumnRuleContract:
     @pytest.mark.parametrize(
         "rule",
         [
-            NotEmptyRule(),
-            MatchesRegexRule(value="^ENSMUSG"),
-            ContainsSubstringRule(value="world"),
-            OneOfRule(value={"a"}),
+            utils.NotEmptyRule(),
+            utils.MatchesRegexRule(value="^ENSMUSG"),
+            utils.ContainsSubstringRule(value="world"),
+            utils.OneOfRule(value={"a"}),
         ],
     )
-    def test_count_violations_returns_int(self, rule: ColumnRule) -> None:
+    def test_count_violations_returns_int(self, rule: utils.ColumnRule) -> None:
         assert isinstance(rule.count_violations(pd.Series(["a", None])), int)
 
     @pytest.mark.parametrize(
         "rule",
         [
-            NotEmptyRule(),
-            MatchesRegexRule(value="^ENSMUSG"),
-            ContainsSubstringRule(value="world"),
-            OneOfRule(value={"a"}),
+            utils.NotEmptyRule(),
+            utils.MatchesRegexRule(value="^ENSMUSG"),
+            utils.ContainsSubstringRule(value="world"),
+            utils.OneOfRule(value={"a"}),
         ],
     )
-    def test_count_violations_is_non_negative(self, rule: ColumnRule) -> None:
+    def test_count_violations_is_non_negative(self, rule: utils.ColumnRule) -> None:
         assert rule.count_violations(pd.Series(["a", None])) >= 0
 
     def test_column_rule_cannot_be_instantiated_directly(self) -> None:
         with pytest.raises(TypeError):
-            ColumnRule()
+            utils.ColumnRule()
 
 
 class TestNotEmptyRule:
@@ -784,34 +779,44 @@ class TestNotEmptyRule:
         return pd.Series(data)
 
     def test_no_violations_for_all_valid(self) -> None:
-        assert NotEmptyRule().count_violations(self._series(["a", "b", "c"])) == 0
+        assert utils.NotEmptyRule().count_violations(self._series(["a", "b", "c"])) == 0
 
     def test_counts_none_as_violation(self) -> None:
-        assert NotEmptyRule().count_violations(self._series(["a", None, "c"])) == 1
+        assert (
+            utils.NotEmptyRule().count_violations(self._series(["a", None, "c"])) == 1
+        )
 
     def test_counts_nan_as_violation(self) -> None:
-        assert NotEmptyRule().count_violations(self._series(["a", np.nan, "c"])) == 1
+        assert (
+            utils.NotEmptyRule().count_violations(self._series(["a", np.nan, "c"])) == 1
+        )
 
     def test_counts_empty_string_as_violation(self) -> None:
-        assert NotEmptyRule().count_violations(self._series(["a", "", "c"])) == 1
+        assert utils.NotEmptyRule().count_violations(self._series(["a", "", "c"])) == 1
 
     def test_counts_whitespace_only_as_violation(self) -> None:
-        assert NotEmptyRule().count_violations(self._series(["a", "   ", "c"])) == 1
+        assert (
+            utils.NotEmptyRule().count_violations(self._series(["a", "   ", "c"])) == 1
+        )
 
     def test_counts_multiple_violations(self) -> None:
         assert (
-            NotEmptyRule().count_violations(self._series(["a", None, "", "   ", "b"]))
+            utils.NotEmptyRule().count_violations(
+                self._series(["a", None, "", "   ", "b"])
+            )
             == 3
         )
 
     def test_all_violations(self) -> None:
-        assert NotEmptyRule().count_violations(self._series([None, "", "   "])) == 3
+        assert (
+            utils.NotEmptyRule().count_violations(self._series([None, "", "   "])) == 3
+        )
 
     def test_empty_series(self) -> None:
-        assert NotEmptyRule().count_violations(self._series([])) == 0
+        assert utils.NotEmptyRule().count_violations(self._series([])) == 0
 
     def test_value_detail_is_empty_string(self) -> None:
-        assert NotEmptyRule().value_detail == ""
+        assert utils.NotEmptyRule().value_detail == ""
 
 
 class TestMatchesRegexRule:
@@ -825,47 +830,47 @@ class TestMatchesRegexRule:
         self, bad_value: int | str | float | None
     ) -> None:
         with pytest.raises(ValueError, match="requires a non-None"):
-            MatchesRegexRule(value=bad_value)
+            utils.MatchesRegexRule(value=bad_value)
 
     def test_raises_when_value_is_invalid_regex(self) -> None:
         with pytest.raises(ValueError, match="valid regex"):
-            MatchesRegexRule(value="[invalid")
+            utils.MatchesRegexRule(value="[invalid")
 
     def test_no_violations_when_all_match(self) -> None:
-        rule = MatchesRegexRule(value="^ENSMUSG")
+        rule = utils.MatchesRegexRule(value="^ENSMUSG")
         assert rule.count_violations(self._series(["ENSMUSG001", "ENSMUSG002"])) == 0
 
     def test_counts_non_matching_value(self) -> None:
-        rule = MatchesRegexRule(value="^ENSMUSG")
+        rule = utils.MatchesRegexRule(value="^ENSMUSG")
         assert rule.count_violations(self._series(["ENSMUSG001", "ENSG002"])) == 1
 
     def test_counts_all_non_matching(self) -> None:
-        rule = MatchesRegexRule(value="^ENSMUSG")
+        rule = utils.MatchesRegexRule(value="^ENSMUSG")
         assert rule.count_violations(self._series(["ENSG001", "ENSG002"])) == 2
 
     def test_skips_none(self) -> None:
         # Nulls are skipped so the rule only validates the format of present values.
-        rule = MatchesRegexRule(value="^ENSMUSG")
+        rule = utils.MatchesRegexRule(value="^ENSMUSG")
         assert rule.count_violations(self._series(["ENSMUSG001", None])) == 0
 
     def test_skips_nan(self) -> None:
-        rule = MatchesRegexRule(value="^ENSMUSG")
+        rule = utils.MatchesRegexRule(value="^ENSMUSG")
         assert rule.count_violations(self._series(["ENSMUSG001", np.nan])) == 0
 
     def test_counts_empty_string_as_violation(self) -> None:
-        rule = MatchesRegexRule(value="^ENSMUSG")
+        rule = utils.MatchesRegexRule(value="^ENSMUSG")
         assert rule.count_violations(self._series(["ENSMUSG001", ""])) == 1
 
     def test_partial_match_is_violation(self) -> None:
-        rule = MatchesRegexRule(value="^ENSMUSG")
+        rule = utils.MatchesRegexRule(value="^ENSMUSG")
         assert rule.count_violations(self._series(["prefix_ENSMUSG001"])) == 1
 
     def test_empty_series(self) -> None:
-        rule = MatchesRegexRule(value="^ENSMUSG")
+        rule = utils.MatchesRegexRule(value="^ENSMUSG")
         assert rule.count_violations(self._series([])) == 0
 
     def test_value_detail_includes_pattern(self) -> None:
-        rule = MatchesRegexRule(value="^ENSMUSG")
+        rule = utils.MatchesRegexRule(value="^ENSMUSG")
         assert "^ENSMUSG" in rule.value_detail
 
 
@@ -878,49 +883,49 @@ class TestContainsSubstringRule:
     @pytest.mark.parametrize("bad_value", [None, np.nan, ""])
     def test_raises_when_value_is_invalid(self, bad_value: str | float | None) -> None:
         with pytest.raises(ValueError, match="requires a non-None"):
-            ContainsSubstringRule(value=bad_value)
+            utils.ContainsSubstringRule(value=bad_value)
 
     def test_no_violations_when_all_contain_substring(self) -> None:
-        rule = ContainsSubstringRule(value="world")
+        rule = utils.ContainsSubstringRule(value="world")
         assert rule.count_violations(self._series(["hello world", "world cup"])) == 0
 
     def test_counts_missing_substring(self) -> None:
-        rule = ContainsSubstringRule(value="world")
+        rule = utils.ContainsSubstringRule(value="world")
         assert rule.count_violations(self._series(["hello world", "goodbye"])) == 1
 
     def test_counts_all_missing(self) -> None:
-        rule = ContainsSubstringRule(value="world")
+        rule = utils.ContainsSubstringRule(value="world")
         assert rule.count_violations(self._series(["foo", "bar"])) == 2
 
     def test_counts_none_as_violation(self) -> None:
-        rule = ContainsSubstringRule(value="world")
+        rule = utils.ContainsSubstringRule(value="world")
         assert rule.count_violations(self._series(["hello world", None])) == 1
 
     def test_counts_nan_as_violation(self) -> None:
-        rule = ContainsSubstringRule(value="world")
+        rule = utils.ContainsSubstringRule(value="world")
         assert rule.count_violations(self._series(["hello world", np.nan])) == 1
 
     def test_counts_empty_string_as_violation(self) -> None:
-        rule = ContainsSubstringRule(value="world")
+        rule = utils.ContainsSubstringRule(value="world")
         assert rule.count_violations(self._series(["hello world", ""])) == 1
 
     def test_counts_non_string_data_as_violations(self) -> None:
-        rule = ContainsSubstringRule(value="world")
+        rule = utils.ContainsSubstringRule(value="world")
         assert (
             rule.count_violations(self._series(["hello world", "goodbye", 2, 5.555555]))
             == 3
         )
 
     def test_value_is_treated_as_literal_not_regex(self) -> None:
-        rule = ContainsSubstringRule(value="-")
+        rule = utils.ContainsSubstringRule(value="-")
         assert rule.count_violations(self._series(["hello-world", "goodbye"])) == 1
 
     def test_empty_series(self) -> None:
-        rule = ContainsSubstringRule(value="world")
+        rule = utils.ContainsSubstringRule(value="world")
         assert rule.count_violations(self._series([])) == 0
 
     def test_value_detail_includes_substring(self) -> None:
-        rule = ContainsSubstringRule(value="world")
+        rule = utils.ContainsSubstringRule(value="world")
         assert "world" in rule.value_detail
 
 
@@ -933,69 +938,69 @@ class TestOneOfRule:
     @pytest.mark.parametrize("bad_value", [None, set(), [], {}])
     def test_raises_when_value_is_invalid(self, bad_value: object) -> None:
         with pytest.raises(ValueError, match="requires a non-None"):
-            OneOfRule(value=bad_value)
+            utils.OneOfRule(value=bad_value)
 
     def test_no_violations_when_all_in_set(self) -> None:
-        rule = OneOfRule(value={"male", "female"})
+        rule = utils.OneOfRule(value={"male", "female"})
         assert rule.count_violations(self._series(["male", "female", "male"])) == 0
 
     def test_counts_value_not_in_set(self) -> None:
-        rule = OneOfRule(value={"male", "female"})
+        rule = utils.OneOfRule(value={"male", "female"})
         assert rule.count_violations(self._series(["male", "unknown"])) == 1
 
     def test_counts_all_invalid(self) -> None:
-        rule = OneOfRule(value={"male", "female"})
+        rule = utils.OneOfRule(value={"male", "female"})
         assert rule.count_violations(self._series(["unknown", "other"])) == 2
 
     def test_counts_none_as_violation(self) -> None:
-        rule = OneOfRule(value={"male", "female"})
+        rule = utils.OneOfRule(value={"male", "female"})
         assert rule.count_violations(self._series(["male", None])) == 1
 
     def test_works_with_list_as_allowed_values(self) -> None:
-        rule = OneOfRule(value=["male", "female"])
+        rule = utils.OneOfRule(value=["male", "female"])
         assert rule.count_violations(self._series(["male", "unknown"])) == 1
 
     def test_empty_series(self) -> None:
-        rule = OneOfRule(value={"male", "female"})
+        rule = utils.OneOfRule(value={"male", "female"})
         assert rule.count_violations(self._series([])) == 0
 
     def test_value_detail_includes_allowed_values(self) -> None:
-        rule = OneOfRule(value={"male"})
+        rule = utils.OneOfRule(value={"male"})
         assert "male" in rule.value_detail
 
     def test_no_violations_numeric_allowed_values(self) -> None:
-        rule = OneOfRule(value=[1, 2, 3])
+        rule = utils.OneOfRule(value=[1, 2, 3])
         assert rule.count_violations(self._series([1, 1, 2])) == 0
 
     def test_no_violations_bool_allowed_values(self) -> None:
-        rule = OneOfRule(value=[True, False])
+        rule = utils.OneOfRule(value=[True, False])
         assert rule.count_violations(self._series([True, False, False])) == 0
 
     def test_no_violations_sentinel_allowed_values(self) -> None:
         # Series.isin matches np.nan to np.nan in the allowed collection (pandas semantics).
-        rule = OneOfRule(value=["", None, [], np.nan])
+        rule = utils.OneOfRule(value=["", None, [], np.nan])
         s = self._series(["", "", None, [], np.nan])
         assert rule.count_violations(s) == 0
 
     def test_no_violations_mixed_int_and_string_allowed(self) -> None:
-        rule = OneOfRule(value=[2, "2"])
+        rule = utils.OneOfRule(value=[2, "2"])
         assert rule.count_violations(self._series([2, 2, "2"])) == 0
 
     def test_violations_when_numeric_allowed_but_string_in_series(self) -> None:
-        rule = OneOfRule(value=[1, 2])
+        rule = utils.OneOfRule(value=[1, 2])
         assert rule.count_violations(self._series([1, 1, "2"])) == 1
 
     def test_violations_all_strings_when_numeric_allowed(self) -> None:
-        rule = OneOfRule(value=[1, 2])
+        rule = utils.OneOfRule(value=[1, 2])
         assert rule.count_violations(self._series(["1", "1", "2"])) == 3
 
     def test_bool_and_int_equivalence_with_bool_allowed(self) -> None:
-        rule = OneOfRule(value=[True, False])
+        rule = utils.OneOfRule(value=[True, False])
         s = self._series([True, 1, False, 0])
         assert rule.count_violations(s) == 0
 
     def test_bool_and_int_equivalence_with_int_allowed(self) -> None:
-        rule = OneOfRule(value=[0, 1])
+        rule = utils.OneOfRule(value=[0, 1])
         s = self._series([True, 1, False, 0])
         assert rule.count_violations(s) == 0
 
@@ -1009,23 +1014,23 @@ class TestCheckColumnRules:
     @pytest.mark.parametrize(
         "col_data, rule",
         [
-            (["a", "b", "c"], NotEmptyRule()),
+            (["a", "b", "c"], utils.NotEmptyRule()),
             (
                 ["ENSMUSG001", "ENSMUSG002"],
-                MatchesRegexRule(value="^ENSMUSG"),
+                utils.MatchesRegexRule(value="^ENSMUSG"),
             ),
             (
                 ["hello world", "world cup"],
-                ContainsSubstringRule(value="world"),
+                utils.ContainsSubstringRule(value="world"),
             ),
             (
                 ["male", "female", "male"],
-                OneOfRule(value={"male", "female"}),
+                utils.OneOfRule(value={"male", "female"}),
             ),
         ],
     )
     def test_rule_passes_for_all_valid_values(
-        self, col_data: list[Any], rule: ColumnRule
+        self, col_data: list[Any], rule: utils.ColumnRule
     ) -> None:
         datasets = self._make_datasets({"col": col_data})
         utils.check_column_rules(datasets, {"ds": {"col": [rule]}})
@@ -1036,45 +1041,45 @@ class TestCheckColumnRules:
     ) -> None:
         datasets = self._make_datasets({"col": ["a", bad_value, "c"]})
         with pytest.raises(ValueError, match="col.*not_empty"):
-            utils.check_column_rules(datasets, {"ds": {"col": [NotEmptyRule()]}})
+            utils.check_column_rules(datasets, {"ds": {"col": [utils.NotEmptyRule()]}})
 
     @pytest.mark.parametrize(
         "col_data, rule, match_pattern",
         [
             (
                 ["ENSMUSG001", "ENSG002", "ENSG003"],
-                MatchesRegexRule(value="^ENSMUSG"),
+                utils.MatchesRegexRule(value="^ENSMUSG"),
                 r"2 row\(s\).*matches_regex.*\^ENSMUSG",
             ),
             (
                 ["hello world", "goodbye"],
-                ContainsSubstringRule(value="world"),
+                utils.ContainsSubstringRule(value="world"),
                 r"1 row\(s\).*contains_substring.*world",
             ),
             (
                 ["hello world", "goodbye", "adieu", "farewell"],
-                ContainsSubstringRule(value="world"),
+                utils.ContainsSubstringRule(value="world"),
                 r"3 row\(s\).*contains_substring.*world",
             ),
             (
                 ["male", "female", "unknown", "other"],
-                OneOfRule(value={"male", "female"}),
+                utils.OneOfRule(value={"male", "female"}),
                 r"2 row\(s\).*one_of",
             ),
             (
                 ["valid", None, ""],
-                NotEmptyRule(),
+                utils.NotEmptyRule(),
                 r"2 row\(s\).*not_empty",
             ),
             (
                 ["valid", None, "", "   ", "also valid"],
-                NotEmptyRule(),
+                utils.NotEmptyRule(),
                 r"3 row\(s\).*not_empty",
             ),
         ],
     )
     def test_rule_raises_with_correct_count(
-        self, col_data: list[Any], rule: ColumnRule, match_pattern: str
+        self, col_data: list[Any], rule: utils.ColumnRule, match_pattern: str
     ) -> None:
         datasets = self._make_datasets({"col": col_data})
         with pytest.raises(ValueError, match=match_pattern):
@@ -1083,11 +1088,11 @@ class TestCheckColumnRules:
     @pytest.mark.parametrize(
         "good_value, rule",
         [
-            ("hello world", ContainsSubstringRule(value="world")),
+            ("hello world", utils.ContainsSubstringRule(value="world")),
         ],
     )
     def test_rule_treats_null_as_violation(
-        self, good_value: str, rule: ColumnRule
+        self, good_value: str, rule: utils.ColumnRule
     ) -> None:
         datasets = self._make_datasets({"col": [good_value, None]})
         with pytest.raises(ValueError, match=rule.rule):
@@ -1097,7 +1102,7 @@ class TestCheckColumnRules:
         # MatchesRegexRule only validates present values; nulls are not violations.
         datasets = self._make_datasets({"col": ["ENSMUSG001", None]})
         utils.check_column_rules(
-            datasets, {"ds": {"col": [MatchesRegexRule(value="^ENSMUSG")]}}
+            datasets, {"ds": {"col": [utils.MatchesRegexRule(value="^ENSMUSG")]}}
         )
 
     def test_all_violations_collected_in_single_error(self) -> None:
@@ -1106,8 +1111,8 @@ class TestCheckColumnRules:
             "ds2": pd.DataFrame({"col_b": ["ENSG001", "ENSG002"]}),
         }
         column_rules = {
-            "ds1": {"col_a": [NotEmptyRule()]},
-            "ds2": {"col_b": [MatchesRegexRule(value="^ENSMUSG")]},
+            "ds1": {"col_a": [utils.NotEmptyRule()]},
+            "ds2": {"col_b": [utils.MatchesRegexRule(value="^ENSMUSG")]},
         }
         with pytest.raises(ValueError) as exc_info:
             utils.check_column_rules(datasets, column_rules)
@@ -1119,7 +1124,7 @@ class TestCheckColumnRules:
         datasets = {"ds": pd.DataFrame({"col": ["a"]})}
         utils.check_column_rules(
             datasets,
-            {"nonexistent_ds": {"col": [NotEmptyRule()]}},
+            {"nonexistent_ds": {"col": [utils.NotEmptyRule()]}},
         )
 
     def test_missing_column_in_rules_reports_violation(self) -> None:
@@ -1127,12 +1132,12 @@ class TestCheckColumnRules:
         with pytest.raises(ValueError, match="does not exist"):
             utils.check_column_rules(
                 datasets,
-                {"ds": {"missing_col": [NotEmptyRule()]}},
+                {"ds": {"missing_col": [utils.NotEmptyRule()]}},
             )
 
     @pytest.mark.parametrize(
         "rule_class",
-        [MatchesRegexRule, ContainsSubstringRule, OneOfRule],
+        [utils.MatchesRegexRule, utils.ContainsSubstringRule, utils.OneOfRule],
     )
     def test_value_required_rule_raises_when_value_is_none(
         self, rule_class: RuleClassWithValueArg
@@ -1774,3 +1779,203 @@ class TestDelimStringToList:
         """
         with pytest.raises(TypeError, match="Delimiter must be a string"):
             utils.delim_string_to_list(input_string, delimiter)
+
+
+class TestRoundYAxisMax:
+    """Test class for the round_y_axis_max function."""
+
+    def test_round_y_axis_max_zero_case(self) -> None:
+        """
+        Test that 0 returns 10.
+
+        Zero is a special case that defaults to 10.0 for visualization purposes.
+        """
+        result = utils.round_y_axis_max(0)
+        assert result == pytest.approx(10.0)
+
+    def test_round_y_axis_max_negative_case(self) -> None:
+        """
+        Test that negative values return 0.
+
+        Negative values are not expected in the data and default to 0.0.
+        """
+        result = utils.round_y_axis_max(-5.0)
+        assert result == pytest.approx(0.0)
+
+    def test_round_y_axis_max_edge_cases(self) -> None:
+        """
+        Test edge cases and boundary conditions.
+
+        Tests very small numbers, "nice" numbers that still need rounding up,
+        and large numbers to ensure the algorithm works across all scales.
+        """
+        # Test very small numbers
+        assert abs(utils.round_y_axis_max(0.0001) - 0.00015) < 1e-6
+        assert abs(utils.round_y_axis_max(0.00001) - 0.000015) < 1e-6
+
+        # Test numbers that are already "nice"
+        assert utils.round_y_axis_max(1.0) == pytest.approx(
+            1.5
+        )  # Should round up to next nice number
+        assert utils.round_y_axis_max(1.5) == pytest.approx(
+            2.0
+        )  # According to JIRA instructions, always round UP to next 5 or 0
+        assert utils.round_y_axis_max(2.0) == pytest.approx(
+            2.5
+        )  # Should round up to next nice number
+
+        # Test large numbers
+        assert utils.round_y_axis_max(1000000) == pytest.approx(1500000)
+        assert utils.round_y_axis_max(5000000) == pytest.approx(
+            5500000
+        )  # 5.0 -> 5.5, not next magnitude
+
+    def test_round_y_axis_max_second_digit_logic(self) -> None:
+        """
+        Test the second digit rounding logic specifically.
+
+        Validates that numbers are always rounded UP to the next value where
+        the second significant digit is 0 or 5.
+        """
+        # Second digit 0-2 should round to 0 (but we round UP, so to 5)
+        assert utils.round_y_axis_max(1.0) == pytest.approx(1.5)
+        assert utils.round_y_axis_max(1.1) == pytest.approx(1.5)
+        assert utils.round_y_axis_max(1.2) == pytest.approx(1.5)
+
+        # Second digit 3-7 should round to 5
+        assert utils.round_y_axis_max(1.3) == pytest.approx(1.5)
+        assert utils.round_y_axis_max(1.4) == pytest.approx(1.5)
+        assert utils.round_y_axis_max(1.5) == pytest.approx(2.0)
+        assert utils.round_y_axis_max(1.6) == pytest.approx(2.0)
+        assert utils.round_y_axis_max(1.7) == pytest.approx(2.0)
+
+        # Second digit 8-9 should round to next first digit with 0
+        assert utils.round_y_axis_max(1.8) == pytest.approx(2.0)
+        assert utils.round_y_axis_max(1.9) == pytest.approx(2.0)
+
+    def test_round_y_axis_max_magnitude_handling(self) -> None:
+        """
+        Test that the function handles different magnitudes correctly.
+
+        Ensures consistent rounding behavior across different orders of magnitude
+        (0.1, 1, 10, 100, 1000, etc.).
+        """
+        # Test different magnitudes with same pattern
+        assert abs(utils.round_y_axis_max(0.1) - 0.15) < 1e-6
+        assert utils.round_y_axis_max(1.0) == pytest.approx(1.5)
+        assert utils.round_y_axis_max(10.0) == pytest.approx(15.0)
+        assert utils.round_y_axis_max(100.0) == pytest.approx(150.0)
+        assert utils.round_y_axis_max(1000.0) == pytest.approx(1500.0)
+
+    def test_round_y_axis_max_floating_point_precision(self) -> None:
+        """
+        Test that floating point precision issues are handled correctly.
+
+        Verifies that common floating point arithmetic issues (like 0.1 + 0.2)
+        don't cause incorrect rounding results.
+        """
+        # Test numbers that might have floating point precision issues
+        assert (
+            abs(utils.round_y_axis_max(0.1 + 0.2) - 0.35) < 1e-6
+        )  # 0.30000000000000004 -> 0.35
+        assert (
+            abs(utils.round_y_axis_max(1.0 / 3.0) - 0.35) < 1e-6
+        )  # 0.3333333333333333 -> 0.35
+        assert (
+            abs(utils.round_y_axis_max(2.0 / 3.0) - 0.7) < 1e-6
+        )  # 0.6666666666666666 -> 0.7
+
+    def test_round_y_axis_max_return_type(self) -> None:
+        """
+        Test that the function returns a float.
+
+        Ensures the return type is always float, even for integer inputs.
+        """
+        result = utils.round_y_axis_max(1.5)
+        assert isinstance(result, float)
+
+    def test_round_y_axis_max_consistency(self) -> None:
+        """
+        Test that the function is consistent across multiple calls.
+
+        Verifies that calling the function multiple times with the same input
+        always produces the same output.
+        """
+        test_values = [0.0021, 1094, 1.616, 0.0, 0.089]
+
+        for value in test_values:
+            result1 = utils.round_y_axis_max(value)
+            result2 = utils.round_y_axis_max(value)
+            assert (
+                result1 == result2
+            ), f"Inconsistent results for {value}: {result1} vs {result2}"
+
+    def test_round_y_axis_max_monotonicity(self) -> None:
+        """
+        Test that the function maintains monotonicity.
+
+        Ensures that larger inputs always produce larger or equal outputs,
+        which is critical for proper data visualization scaling.
+        """
+        test_values = [0.001, 0.002, 0.003, 0.004, 0.005]
+        results = [utils.round_y_axis_max(val) for val in test_values]
+
+        # Results should be non-decreasing
+        for i in range(1, len(results)):
+            assert (
+                results[i] >= results[i - 1]
+            ), f"Non-monotonic: {test_values[i-1]} -> {results[i-1]}, {test_values[i]} -> {results[i]}"
+
+    @pytest.mark.parametrize(
+        "input_val,expected",
+        [
+            (0, 10.0),
+            (-1, 0.0),
+            (0.0021, 0.0025),
+            (0.0004, 0.00045),
+            (0.329486078, 0.35),
+            (0.089, 0.090),
+            (1094, 1500),
+            (1322498, 1500000),
+            (728591, 750000),
+            (3973, 4000),
+            (1.616, 2.0),
+        ],
+    )
+    def test_round_y_axis_max_parametrized(
+        self, input_val: float, expected: float
+    ) -> None:
+        """
+        Parametrized test for the main examples.
+
+        Comprehensive parametrized test covering all key test cases including
+        zero, negative, and various positive values from the JIRA specification.
+
+        Args:
+            input_val: Input value to test
+            expected: Expected rounded output
+        """
+        result = utils.round_y_axis_max(input_val)
+        assert (
+            abs(result - expected) < 1e-6
+        ), f"input={input_val}, expected={expected}, got={result}"
+
+    def test_round_y_axis_max_string_inputs(self) -> None:
+        """
+        Test that string inputs are properly converted and handled.
+
+        Validates that the function accepts string inputs (per type signature)
+        and properly converts valid numeric strings or returns 10.0 for invalid ones.
+        """
+        # Valid string numbers
+        assert abs(utils.round_y_axis_max("1.5") - 2.0) < 1e-6
+        assert abs(utils.round_y_axis_max("100") - 150.0) < 1e-6
+        assert abs(utils.round_y_axis_max("0.0021") - 0.0025) < 1e-6
+
+        # Invalid string (should return 10.0)
+        assert utils.round_y_axis_max("invalid") == pytest.approx(10.0)
+        assert utils.round_y_axis_max("abc123") == pytest.approx(10.0)
+        assert utils.round_y_axis_max("") == pytest.approx(10.0)
+
+        # None input (should return 10.0)
+        assert utils.round_y_axis_max(None) == pytest.approx(10.0)
