+++ tests/transform/test_immunohisto_transform.py
@@ -21,7 +21,6 @@ from agoradatatools.etl.transform.immunohisto_transform import (
     _extract_age_num,
     immunohisto_transform,
     prepare_immunohisto_data,
-    round_y_axis_max,
 )
 
 
@@ -548,204 +547,8 @@ class TestTransformGeneralModelAD:
         assert evidence_types == expected_order
 
 
-class TestRoundYAxisMax:
-    """Test class for the round_y_axis_max function."""
-
-    def test_round_y_axis_max_zero_case(self) -> None:
-        """
-        Test that 0 returns 10.
-
-        Zero is a special case that defaults to 10.0 for visualization purposes.
-        """
-        result = round_y_axis_max(0)
-        assert result == pytest.approx(10.0)
-
-    def test_round_y_axis_max_negative_case(self) -> None:
-        """
-        Test that negative values return 0.
-
-        Negative values are not expected in the data and default to 0.0.
-        """
-        result = round_y_axis_max(-5.0)
-        assert result == pytest.approx(0.0)
-
-    def test_round_y_axis_max_edge_cases(self) -> None:
-        """
-        Test edge cases and boundary conditions.
-
-        Tests very small numbers, "nice" numbers that still need rounding up,
-        and large numbers to ensure the algorithm works across all scales.
-        """
-        # Test very small numbers
-        assert abs(round_y_axis_max(0.0001) - 0.00015) < 1e-6
-        assert abs(round_y_axis_max(0.00001) - 0.000015) < 1e-6
-
-        # Test numbers that are already "nice"
-        assert round_y_axis_max(1.0) == pytest.approx(
-            1.5
-        )  # Should round up to next nice number
-        assert round_y_axis_max(1.5) == pytest.approx(
-            2.0
-        )  # According to JIRA instructions, always round UP to next 5 or 0
-        assert round_y_axis_max(2.0) == pytest.approx(
-            2.5
-        )  # Should round up to next nice number
-
-        # Test large numbers
-        assert round_y_axis_max(1000000) == pytest.approx(1500000)
-        assert round_y_axis_max(5000000) == pytest.approx(
-            5500000
-        )  # 5.0 -> 5.5, not next magnitude
-
-    def test_round_y_axis_max_second_digit_logic(self) -> None:
-        """
-        Test the second digit rounding logic specifically.
-
-        Validates that numbers are always rounded UP to the next value where
-        the second significant digit is 0 or 5.
-        """
-        # Second digit 0-2 should round to 0 (but we round UP, so to 5)
-        assert round_y_axis_max(1.0) == pytest.approx(1.5)
-        assert round_y_axis_max(1.1) == pytest.approx(1.5)
-        assert round_y_axis_max(1.2) == pytest.approx(1.5)
-
-        # Second digit 3-7 should round to 5
-        assert round_y_axis_max(1.3) == pytest.approx(1.5)
-        assert round_y_axis_max(1.4) == pytest.approx(1.5)
-        assert round_y_axis_max(1.5) == pytest.approx(2.0)
-        assert round_y_axis_max(1.6) == pytest.approx(2.0)
-        assert round_y_axis_max(1.7) == pytest.approx(2.0)
-
-        # Second digit 8-9 should round to next first digit with 0
-        assert round_y_axis_max(1.8) == pytest.approx(2.0)
-        assert round_y_axis_max(1.9) == pytest.approx(2.0)
-
-    def test_round_y_axis_max_magnitude_handling(self) -> None:
-        """
-        Test that the function handles different magnitudes correctly.
-
-        Ensures consistent rounding behavior across different orders of magnitude
-        (0.1, 1, 10, 100, 1000, etc.).
-        """
-        # Test different magnitudes with same pattern
-        assert abs(round_y_axis_max(0.1) - 0.15) < 1e-6
-        assert round_y_axis_max(1.0) == pytest.approx(1.5)
-        assert round_y_axis_max(10.0) == pytest.approx(15.0)
-        assert round_y_axis_max(100.0) == pytest.approx(150.0)
-        assert round_y_axis_max(1000.0) == pytest.approx(1500.0)
-
-    def test_round_y_axis_max_floating_point_precision(self) -> None:
-        """
-        Test that floating point precision issues are handled correctly.
-
-        Verifies that common floating point arithmetic issues (like 0.1 + 0.2)
-        don't cause incorrect rounding results.
-        """
-        # Test numbers that might have floating point precision issues
-        assert (
-            abs(round_y_axis_max(0.1 + 0.2) - 0.35) < 1e-6
-        )  # 0.30000000000000004 -> 0.35
-        assert (
-            abs(round_y_axis_max(1.0 / 3.0) - 0.35) < 1e-6
-        )  # 0.3333333333333333 -> 0.35
-        assert (
-            abs(round_y_axis_max(2.0 / 3.0) - 0.7) < 1e-6
-        )  # 0.6666666666666666 -> 0.7
-
-    def test_round_y_axis_max_return_type(self) -> None:
-        """
-        Test that the function returns a float.
-
-        Ensures the return type is always float, even for integer inputs.
-        """
-        result = round_y_axis_max(1.5)
-        assert isinstance(result, float)
-
-    def test_round_y_axis_max_consistency(self) -> None:
-        """
-        Test that the function is consistent across multiple calls.
-
-        Verifies that calling the function multiple times with the same input
-        always produces the same output.
-        """
-        test_values = [0.0021, 1094, 1.616, 0.0, 0.089]
-
-        for value in test_values:
-            result1 = round_y_axis_max(value)
-            result2 = round_y_axis_max(value)
-            assert (
-                result1 == result2
-            ), f"Inconsistent results for {value}: {result1} vs {result2}"
-
-    def test_round_y_axis_max_monotonicity(self) -> None:
-        """
-        Test that the function maintains monotonicity.
-
-        Ensures that larger inputs always produce larger or equal outputs,
-        which is critical for proper data visualization scaling.
-        """
-        test_values = [0.001, 0.002, 0.003, 0.004, 0.005]
-        results = [round_y_axis_max(val) for val in test_values]
-
-        # Results should be non-decreasing
-        for i in range(1, len(results)):
-            assert (
-                results[i] >= results[i - 1]
-            ), f"Non-monotonic: {test_values[i-1]} -> {results[i-1]}, {test_values[i]} -> {results[i]}"
-
-    @pytest.mark.parametrize(
-        "input_val,expected",
-        [
-            (0, 10.0),
-            (-1, 0.0),
-            (0.0021, 0.0025),
-            (0.0004, 0.00045),
-            (0.329486078, 0.35),
-            (0.089, 0.090),
-            (1094, 1500),
-            (1322498, 1500000),
-            (728591, 750000),
-            (3973, 4000),
-            (1.616, 2.0),
-        ],
-    )
-    def test_round_y_axis_max_parametrized(
-        self, input_val: float, expected: float
-    ) -> None:
-        """
-        Parametrized test for the main examples.
-
-        Comprehensive parametrized test covering all key test cases including
-        zero, negative, and various positive values from the JIRA specification.
-
-        Args:
-            input_val: Input value to test
-            expected: Expected rounded output
-        """
-        result = round_y_axis_max(input_val)
-        assert (
-            abs(result - expected) < 1e-6
-        ), f"input={input_val}, expected={expected}, got={result}"
-
-    def test_round_y_axis_max_string_inputs(self) -> None:
-        """
-        Test that string inputs are properly converted and handled.
-
-        Validates that the function accepts string inputs (per type signature)
-        and properly converts valid numeric strings or returns 10.0 for invalid ones.
-        """
-        # Valid string numbers
-        assert abs(round_y_axis_max("1.5") - 2.0) < 1e-6
-        assert abs(round_y_axis_max("100") - 150.0) < 1e-6
-        assert abs(round_y_axis_max("0.0021") - 0.0025) < 1e-6
-
-        # Invalid string (should return 10.0)
-        assert round_y_axis_max("invalid") == pytest.approx(10.0)
-        assert round_y_axis_max("abc123") == pytest.approx(10.0)
-        assert round_y_axis_max("") == pytest.approx(10.0)
-
-        # None input (should return 10.0)
-        assert round_y_axis_max(None) == pytest.approx(10.0)
+class TestImmunohistoTransformInvalidAge:
+    """Test class for immunohisto_transform's handling of invalid age values."""
 
     def test_extract_age_num_valueerror_handling(self) -> None:
         """Test that non-numeric age values raise a descriptive error."""
